requirements.txt
tesseract
pytesseract
opencv-python
Pillow
Flask-Uploads
PyMuPDF
Flask-Reuploaded
convertapi